# Conflicts:
#	src/MEGASync/MegaApplication.cpp
#	src/MEGASync/MegaApplication.h
#	src/MEGASync/control/Utilities.cpp
#	src/MEGASync/control/Utilities.h
#	src/MEGASync/gui/InfoDialog.cpp
#	src/MEGASync/gui/SettingsDialog.cpp
#	src/MEGASync/gui/ViewLoadingScene.h
#	src/MEGASync/gui/node_selector/gui/NodeSelector.cpp
#	src/MEGASync/gui/node_selector/gui/NodeSelectorTreeView.cpp
#	src/MEGASync/gui/node_selector/gui/NodeSelectorTreeView.h
#	src/MEGASync/gui/node_selector/gui/NodeSelectorTreeViewWidget.cpp
#	src/MEGASync/gui/node_selector/gui/NodeSelectorTreeViewWidget.h
#	src/MEGASync/gui/node_selector/model/NodeSelectorModel.cpp
#	src/MEGASync/gui/node_selector/model/NodeSelectorModel.h
#	src/MEGASync/gui/translations/translation.source.ts
#	src/MEGASync/stalled_issues/model/StalledIssuesUtilities.cpp

+++ src/MEGASync/gui/LowDiskSpaceDialog.cpp
@@ -6,11 +6,13 @@
 
 #include <QFileIconProvider>
 
-LowDiskSpaceDialog::LowDiskSpaceDialog(qint64 neededSize, qint64 freeSize, qint64 driveSize,
-                                       const DriveDisplayData &driveDisplayData,
-                                       QWidget *parent) :
-    QDialog(parent),
-    ui(new Ui::LowDiskSpaceDialog)
+LowDiskSpaceDialog::LowDiskSpaceDialog(long neededSize,
+    long long freeSize,
+    long long driveSize,
+    const DriveDisplayData& driveDisplayData,
+    QWidget* parent)
+    : QDialog(parent)
+    , ui(new Ui::LowDiskSpaceDialog)
 {
     ui->setupUi(this);
     setupShadowEffect();
@@ -34,7 +36,7 @@ LowDiskSpaceDialog::~LowDiskSpaceDialog()
     delete ui;
 }
 
-QString LowDiskSpaceDialog::toString(qint64 bytes)
+QString LowDiskSpaceDialog::toString(long long bytes)
 {
     return Utilities::getSizeString((bytes > 0) ? bytes : 0);
 }
