+++ src/MEGASync/gui/QSegmentedControl.h
@@ -24,7 +24,7 @@ public:
     QSegmentedControl(QWidget *parent = 0);
     void configureTableSegment();
     void configureTabSegment(QStringList options);
-    void clicked(int segment);
+    void clicked(long segment);
     void setTableButtonEnable(int segment, bool enable);
     ~QSegmentedControl();
