+++ src/MEGASync/gui/node_selector/model/NodeSelectorModelItem.h
@@ -74,7 +74,7 @@ protected:
     QString mOwnerEmail;
     Status mStatus;
     bool mRequestingChildren;
-    long long mChildrenCounter;
+    int mChildrenCounter;
     bool mShowFiles;
     bool mChildrenAreInit;
